+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 4.71.2 (2026-02-02)
+
 - fixed: Remove extra call to whitelistAssets that was hanging and preventing core callbacks from resolving
 
 ## 4.71.1 (2026-01-28)